+++ CHANGELOG.md
@@ -14,6 +14,20 @@ On the bright side we also aren't simply going away. There is a certain secret p
 
 # BCX Changelog
 
+## 1.1.4
+
+This update changes the following:
+- Switched to `ChatRoomAppendChat()` for appending BCX messages to the chat log
+- Updated dependencies to BC R116
+- Updated backgrounds and fixed missing text errors in the background viewer
+- Made hidden backgrounds only appear in their own tag and not the "all tags" list
+
+## 1.1.3
+
+This update adds the following:
+- Added a new "Farewell on slow leave" rule that makes the affected player say a specific message when one of the room's guests leaves the room (only slow leaves).
+- Improved the "Greet new guests" rule to allow customization of the lowest role that one should greet.
+
 ## 1.1.2
 
 This update restores a few of the rules removed in the earlier versions: